Rename AppStorePanel file, add Query Param Logic
