+++ src/Components/AppStore/AppStorePanelControl.jsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, {useEffect} from 'react'
 import Box from '@mui/material/Box'
 import {BackButton, CloseButton, FullScreenButton} from '../Buttons'
 import useStore from '../../store/useStore'
@@ -9,6 +9,19 @@ import {PanelWithTitle} from '../SideDrawer/SideDrawerPanels'
 /** @return {React.Component} */
 export function AppStorePanel() {
   const toggleAppStoreDrawer = useStore((state) => state.toggleAppStoreDrawer)
+  const setIsAppStoreEnabled = useStore((state) => state.setIsAppStoreEnabled)
+
+  /**
+   * Store initial query parameters settings,
+   * since they will be cleared by the application on state change.
+   */
+  useEffect(() => {
+    const initialParameters = new URLSearchParams(window.location.search)
+    const enabledFeature = initialParameters.get('feature')
+    const appStoreEnabled = enabledFeature && enabledFeature.toLowerCase() === 'apps'
+    setIsAppStoreEnabled(appStoreEnabled)
+  }, [setIsAppStoreEnabled])
+
   return (
     <PanelWithTitle title={'App Store'}
       controlsGroup={
